% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_text_twowaycount.R
\name{compute_group_twowaycount}
\alias{compute_group_twowaycount}
\title{Title}
\usage{
compute_group_twowaycount(data, scales)
}
\arguments{
\item{scales}{}
}
\value{

}
\description{
Title
}
\examples{
# step 1b test the computation function
library(dplyr)
mtcars \%>\%
  # input must have required aesthetic inputs as columns
  rename(x = cyl, y = gear) \%>\%
  compute_group_twowaycount() \%>\%
  head()
}
