% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tile_prop_within_x.R
\name{compute_panel_prop_withinx}
\alias{compute_panel_prop_withinx}
\title{Title}
\usage{
compute_panel_prop_withinx(data, scales, drop = F)
}
\arguments{
\item{scales}{}
}
\value{

}
\description{
Title
}
\examples{
mtcars \%>\%
rename(x = cyl, y = gear, value = 1) \%>\%
compute_panel_prop_withinx
}
