% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_text_prop_within_x.R
\name{geom_text_prop_within_x}
\alias{geom_text_prop_within_x}
\title{Title}
\usage{
geom_text_prop_within_x(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
library(ggplot2)
ggplot(data = mtcars) +
aes(x = cyl, y = gear) +
geom_tile_prop_within_x() +
aes(fill = 1) +
geom_text_prop_within_x(color = "oldlace")
}
