% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_text_twowaycount.R
\name{geom_text_twowaycount}
\alias{geom_text_twowaycount}
\title{Title}
\usage{
geom_text_twowaycount(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
library(ggplot2)
ggplot(data = mtcars) +
aes(x = cyl, y = gear) +
geom_text_twowaycount()
}
