% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tile_prop_within_x.R
\name{geom_tile_prop_withinx}
\alias{geom_tile_prop_withinx}
\title{Title}
\usage{
geom_tile_prop_withinx(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
library(ggplot2)
ggplot(data = mtcars) +
aes(x = cyl, y = gear) +
geom_tile_prop_withinx() +
aes(value = 1, within = cyl) +
geom_text_twowaycount(color = "oldlace")
}
